Do not delete dynamic IP addresses when updating/deleting devices

+++ lib/Netdot/Model/Device.pm
@@ -1831,7 +1831,6 @@ sub fwt_update {
     None
   Returns:
     True if successful
-
   Examples:
     $device->delete();
 
@@ -1840,15 +1839,25 @@ sub delete {
     my ($self) = @_;
     $self->isa_object_method('delete');
 
+    # We don't want to delete dynamic addresses
+    if ( my $ips = $self->get_ips ){
+	foreach my $ip ( @$ips ) {
+	    if ( $ip->status && $ip->status->name eq 'Dynamic' ){
+		$ip->update({interface=>0});
+	    }
+	}
+    }
+
+    # If the RR had a RRADDR, it was probably deleted.  
+    # Otherwise, we do it here.
     my $rrid = ( $self->name )? $self->name->id : "";
     
     $self->SUPER::delete();
 
-    # If the RR had a RRADDR, it was probably deleted.  
-    # Otherwise, we do it here.
     if ( my $rr = RR->retrieve($rrid) ){
 	$rr->delete() unless $rr->arecords;
     }
+    return 1;
 }
 
 ############################################################################
@@ -4797,7 +4806,10 @@ sub _update_interfaces {
 	# (could have been deleted if its interface was deleted)
 	next unless ( defined $obj );
 	next if ( ref($obj) =~ /deleted/i );
-	
+
+	# Leave dynamic addresses alone
+	next if ( $obj->status && $obj->status->name eq 'Dynamic' );
+
 	$logger->info(sprintf("%s: IP %s no longer exists.  Removing.", 
 			      $host, $obj->address));
 	$obj->delete(no_update_tree=>1);
